// frontend/inerface/onlineOrderForm.interface.ts
export interface FormValuesInterface {
    name: string;
    phone: string;
    email: string;
    street: string;
    house: string;
    floor: string;
    apartment: string;
    entrance: string;
    intercom: string;
    testDay: boolean;
    agree: boolean;
    weeks: number;
    deliveryTime: string;
    paymentMethod: string;
    contact: string;
    wishes?: string;
    needCutlery: boolean;
}